day8/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "day8-ai-chat"
version = "0.1.0"
description = "Simple AI chat application using LangChain with z.ai GLM API"
authors = [
    {name = "Elite Engineers", email = "team@elite-engineers.com"}
]
dependencies = [
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langchain>=1.0.7",
    "langchain-openai>=1.0.3",
    "langchain-core>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.1.13",
    "langchain-huggingface>=0.1.0",
    "langchain-text-splitters>=0.1.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "fastmcp>=2.13.1",
    "sentence-transformers>=2.2.2",
    "pypdf>=3.17.4",
    "pdfplumber>=0.10.3",
    "unstructured[pdf]>=0.14.0",
    "openai-whisper>=20250625",
    "numpy>=1.24.0",
    "click>=8.1.0",
    "sounddevice>=0.5.1",
    "soundfile>=0.13.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "black>=24.10.0",
    "ruff>=0.8.6",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']
